0.6

+++ app/Http/Controllers/userController.php
@@ -32,6 +32,10 @@ class userController extends Controller
      public function userdetail($name){
 
         $user = User::where('ad', '=', $name)->first();
+        if(!$user)
+        {
+           abort(404);
+        }
         return view('website/profil', compact('user'));
 
  
@@ -47,14 +51,20 @@ class userController extends Controller
 
    }
    public function useredit(Request $request, User $user){
-      $image = $request->file('image');
-      $input['imagename'] = time() . '.' . $image->getClientOriginalExtension();
-      $path = public_path('/images');
-      $image->move($path,$input['imagename']);
-      File::delete('images/' . $user->img);
+      if($request->file('image'))
+      {
+         $image = $request->file('image');
+         $input['imagename'] = time() . '.' . $image->getClientOriginalExtension();
+         $path = public_path('/images');
+         $image->move($path,$input['imagename']);
+         File::delete('images/' . $user->img);
+         $user->img=$input['imagename'];
+      }
+      if ($request->ad=="") {
+         $request->ad=$user->ad;
+      }
       $user->ad=$request->ad;
       $user->bio=$request->bio;
-      $user->img=$input['imagename'];
       $user->save();
       Session::forget('user');
       return redirect('/1');
